moved gest/set point logic from Increment/Increment.cc to Fields/Fields.cc
